+++ Dashboard/UserCardView.m
@@ -23,9 +23,9 @@
 - (instancetype)initWithImageURL:(NSURL *)imageURL
                             name:(NSString *)name
                        cityState:(NSString *)cityState
-                      emailCount:(int)emailCount
-                        smsCount:(int)smsCount
-                      phoneCount:(int)phoneCount {
+                      emailCount:(NSInteger)emailCount
+                        smsCount:(NSInteger)smsCount
+                      phoneCount:(NSInteger)phoneCount {
     
     if ((self = [super init])) {
         self = [[[NSBundle mainBundle] loadNibNamed:@"UserCardView"
